Added support for SQLAlchemy, improved config support
