+++ Pipfile
@@ -24,4 +24,4 @@ isort = "*"
 pytest = "*"
 
 [requires]
-python_version = "3.11"
+python_version = "3.10"
