+++ settings.conf
@@ -1,3 +1,12 @@
+[main]
+db_type: sqlite
+db_path: db.sqlite
+#db_host: localhost
+#db_port: 3306
+#db_username: None
+#db_password: None
+#db_name: csp_reports
+
 [mongodb]
 enable = False
 host = localhost
@@ -5,4 +14,3 @@ port = 27017
 user = None
 pass = None
 database = csp_reports
-
